refactored function_app code to use newly refactored functions
